+++ test/spec/xform.spec.js
@@ -15,6 +15,12 @@ describe( 'XForm', () => {
             const result = await validator.validate( xf );
             expect( result.duration ).to.be.above( 0 );
         } );
+
+        it( 'returns no errors and no warnings', async() => {
+            const result = await validator.validate( xf );
+            expect( result.errors.length ).to.equal( 0 );
+            expect( result.warnings.length ).to.equal( 0 );
+        } );
     } );
 
     describe( 'that is invalid', () => {
@@ -150,7 +156,7 @@ describe( 'XForm', () => {
     describe( 'with calculations on a form control that are not set to readonly', () => {
         const validation = validator.validate( loadXForm( 'calculation-not-readonly.xml' ) );
 
-        it( 'outputs errors', async() => {
+        it( 'returns errors', async() => {
             const result = await validation;
             expect( arrContains( result.errors, /"a" has a calculation that is not set to readonly/i ) ).to.equal( true );
         } );
@@ -163,12 +169,12 @@ describe( 'XForm', () => {
                 openclinica: true
             } );
 
-            it( 'outputs errors', async() => {
+            it( 'returns errors', async() => {
                 const result = await validation;
                 expect( result.errors.length ).to.equal( 6 );
             } );
 
-            it( 'outputs errors for calculations without form control that refer to external ' +
+            it( 'returns errors for calculations without form control that refer to external ' +
                 'clinicaldata instance but do not have the oc:external="clinicaldata" bind', async() => {
                 const result = await validation;
                 expect( arrContains( result.errors, /"invalid1" .* to external clinicaldata without the required "external" attribute/i ) ).to.equal( true );
@@ -176,7 +182,7 @@ describe( 'XForm', () => {
                 expect( arrContains( result.errors, /"invalid3" .* to external clinicaldata without the required "external" attribute/i ) ).to.equal( true );
             } );
 
-            it( 'outputs errors for binds with oc:external="clinicaldata" that do not ' +
+            it( 'returns errors for binds with oc:external="clinicaldata" that do not ' +
                 'do not have a calculation that refers to instance(\'clinicaldata\')', async() => {
                 const result = await validation;
                 expect( arrContains( result.errors, /"invalid4" .* not .* calculation referring to instance\("clinicaldata"\)/i ) ).to.equal( true );
@@ -204,7 +210,7 @@ describe( 'XForm', () => {
                 openclinica: true
             } );
 
-            it( 'outputs errors', async() => {
+            it( 'returns errors', async() => {
                 const result = await validation;
                 expect( result.errors.length ).to.equal( 9 );
                 expect( arrContains( result.errors, /unsupported oc:constraint .+ "something"/ ) ).to.equal( true );
@@ -224,7 +230,7 @@ describe( 'XForm', () => {
                 openclinica: true
             } );
 
-            it( 'outputs an error', async() => {
+            it( 'returns an error', async() => {
                 const result = await validation;
                 expect( result.errors.length ).to.equal( 1 );
                 expect( arrContains( result.errors, /last-saved\s+not supported/ ) );
@@ -237,7 +243,7 @@ describe( 'XForm', () => {
     describe( 'forms using the special last-saved instance', () => {
         const validation = validator.validate( loadXForm( 'last-saved.xml' ) );
 
-        it( 'does not output an error', async() => {
+        it( 'does not return an error', async() => {
             const result = await validation;
             expect( result.errors.length ).to.equal( 0 );
         } );
@@ -252,7 +258,7 @@ describe( 'XForm', () => {
         const WARNINGS = 14;
         const ERRORS = 1;
 
-        it( 'outputs warnings', async() => {
+        it( 'returns warnings', async() => {
             const result = await validation;
 
             expect( result.warnings.length ).to.equal( WARNINGS );
@@ -272,25 +278,25 @@ describe( 'XForm', () => {
             expect( arrContains( result.warnings, /"field-list" for "two"/i ) ).to.equal( true );
         } );
 
-        it( 'outputs 1 error', async() => {
+        it( 'returns 1 error', async() => {
             const result = await validation;
             expect( result.errors.length ).to.equal( ERRORS );
             expect( arrContains( result.errors, /"search" for question "l"/i ) ).to.equal( true );
         } );
 
-        it( 'outputs 1 error with --oc flag', async() => {
+        it( 'returns 1 error with --oc flag', async() => {
             const resultOc = await validationOc;
             expect( resultOc.errors.length ).to.equal( ERRORS );
             expect( arrContains( resultOc.errors, /"search" for question "l"/i ) ).to.equal( true );
         } );
 
-        it( 'outputs warnings with --oc flag too', async() => {
+        it( 'returns warnings with --oc flag too', async() => {
             const resultOc = await validationOc;
             //expect( arrContains( result.warnings, /deprecated/ ) ).to.equal( false );
             expect( resultOc.warnings.length ).to.equal( WARNINGS );
         } );
 
-        it( 'including the special case "horizontal" output warnings', async() => {
+        it( 'including the special case "horizontal" return warnings', async() => {
             const result = await validator.validate( loadXForm( 'appearance-horizontal.xml' ) );
 
             expect( result.warnings.length ).to.equal( 4 );
@@ -314,7 +320,7 @@ describe( 'XForm', () => {
         const validation = validator.validate( xf );
         const WARNINGS = 3;
 
-        it( 'outputs warnings', async() => {
+        it( 'returns warnings', async() => {
             const result = await validation;
 
             expect( result.warnings.length ).to.equal( WARNINGS );
@@ -328,7 +334,7 @@ describe( 'XForm', () => {
         const xf = loadXForm( 'user-ues.xml' );
         const validation = validator.validate( xf );
 
-        it( 'outputs warnings', async() => {
+        it( 'returns warnings', async() => {
             const result = await validation;
 
             expect( result.warnings.length ).to.equal( 12 );
@@ -351,7 +357,7 @@ describe( 'XForm', () => {
         const xf = loadXForm( 'external-app.xml' );
         const validation = validator.validate( xf );
 
-        it( 'outputs warnings', async() => {
+        it( 'returns warnings', async() => {
             const result = await validation;
 
             expect( result.errors.length ).to.equal( 2 );
@@ -363,7 +369,7 @@ describe( 'XForm', () => {
 
     describe( 'with missing <label> elements', () => {
 
-        it( 'outputs errors', async() => {
+        it( 'returns errors', async() => {
             const result = await validator.validate( loadXForm( 'missing-labels.xml' ) );
             const ISSUES = 6;
             expect( result.errors.length ).to.equal( ISSUES );
@@ -375,7 +381,22 @@ describe( 'XForm', () => {
             expect( arrContains( result.errors, /option for question "i" has no label/i ) ).to.equal( true );
         } );
 
-        it( 'does not output errors for setvalue actions without a label', async() => {
+        it( 'does not return errors for setvalue actions without a label', async() => {
+            const result = await validator.validate( loadXForm( 'setvalue.xml' ) );
+            expect( result.errors.length ).to.equal( 0 );
+        } );
+
+    } );
+
+    describe( 'with missing <value> elements', () => {
+
+        it( 'returns errors', async() => {
+            const result = await validator.validate( loadXForm( 'missing-values.xml' ) );
+            expect( result.errors.length ).to.equal( 1 );
+            expect( arrContains( result.errors, /option for question "k" has no value/i ) ).to.equal( true );
+        } );
+
+        it( 'does not return errors for setvalue actions without a label', async() => {
             const result = await validator.validate( loadXForm( 'setvalue.xml' ) );
             expect( result.errors.length ).to.equal( 0 );
         } );
@@ -384,7 +405,7 @@ describe( 'XForm', () => {
 
     describe( 'with duplicate nodenames', () => {
 
-        it( 'outputs warnings', async() => {
+        it( 'returns warnings', async() => {
             const result = await validator.validate( loadXForm( 'duplicate-nodename.xml' ) );
             expect( result.warnings.length ).to.equal( 2 );
             expect( arrContains( result.warnings, /Duplicate .* name "a" found/i ) ).to.equal( true );
@@ -396,7 +417,7 @@ describe( 'XForm', () => {
 
     describe( 'with nodenames containing underscores', () => {
 
-        it( 'outputs warnings', async() => {
+        it( 'returns warnings', async() => {
             const result = await validator.validate( loadXForm( 'nodename-underscore.xml' ) );
             expect( result.warnings.length ).to.equal( 0 );
             expect( result.errors.length ).to.equal( 0 );
@@ -406,7 +427,7 @@ describe( 'XForm', () => {
 
     describe( 'with nested repeats', () => {
 
-        it( 'outputs warnings', async() => {
+        it( 'returns warnings', async() => {
             const result = await validator.validate( loadXForm( 'nested-repeats.xml' ) );
             expect( result.warnings.length ).to.equal( 2 );
             expect( arrContains( result.warnings, /Repeat "immunization-info" .* nested/i ) ).to.equal( true );
@@ -417,7 +438,7 @@ describe( 'XForm', () => {
 
     xdescribe( 'with disallowed self-referencing', () => {
 
-        it( 'outputs errors for disallowed self-referencing', async() => {
+        it( 'returns errors for disallowed self-referencing', async() => {
             // Unit tests are in xpath.spec.js
             const result = await validator.validate( loadXForm( 'self-reference.xml' ) );
             expect( result.errors.length ).to.equal( 2 );