+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+--------------------------
+##### Added
+- Check for missing value elements in choice options.
+
 [2.0.1] - 2022-11-07
 --------------------------
 ##### Changed